Add files via upload
This project is a web-based calculator designed using HTML, CSS, and JavaScript. It allows users to perform basic arithmetic operations, such as addition, subtraction, multiplication, and division. The calculator features a clean and intuitive user interface, making it easy to use on various devices. By uploading this project to GitHub, it serves as a demonstration of front-end development skills and provides a valuable resource for those interested in learning about web-based calculator implementation. Users can explore the code, contribute improvements, and use it as a foundation for similar projects in the future.